requirements.txt
mcp>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
googlemaps>=4.10.0
python-dateutil>=2.8.2
pydantic>=2.5.0
typing-extensions>=4.8.0
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0 